fix clash/error at preprocessing time due to the included magic_enum header file invoking the assert macros itself: we're looking at a definition loop.
The idea here was to make the preprocessor pahse stand out and come first (as it would otherwise) WITHOUT already spitting out the libassert C++ namespace support code. This concept works (by way of the assert-macros.hpp driver) but then the trouble (at compile time) moves on: the macro expansion in magic_enum header file, now done properly itself, expands into code using expression decomposition and other techniques which depend on the C++ libassert codebase being present in full, which CANNOT happen at that moment in (compile) time as magic_enum support code MUST come BEFORE the libassert support code, as the latter employs the former.

So this idea was implemented and tested OK, but failed to deliver, so we'll be reverting this bugger right next.....